# docs/gen/config/agent/config-swagger.yaml
basePath: /aperture-agent
definitions:
...
    ClientTLSConfig:
        properties:
            ca_file:
                type: string
                x-go-name: CAFile
                x-go-tag-json: ca_file
            cert_file:
                type: string
                x-go-name: CertFile
                x-go-tag-json: cert_file
            insecure_skip_verify:
                type: boolean
                x-go-name: InsecureSkipVerify
                x-go-tag-json: insecure_skip_verify
            key_file:
                type: string
                x-go-name: KeyFile
                x-go-tag-json: key_file
            key_log_file:
                type: string
                x-go-name: KeyLogWriter
                x-go-tag-json: key_log_file
        title: ClientTLSConfig is the configuration for client TLS.
        type: object
        x-go-package: github.com/fluxninja/aperture/v2/pkg/net/tlsconfig
...
    HTTPClientConfig:
        properties:
            disable_compression:
                default: false
                description: Disable Compression
                type: boolean
                x-go-name: DisableCompression
                x-go-tag-default: "false"
                x-go-tag-json: disable_compression
            disable_keep_alives:
                default: false
                description: Disable HTTP Keepalive
                type: boolean
                x-go-name: DisableKeepAlives
                x-go-tag-default: "false"
                x-go-tag-json: disable_keep_alives
            expect_continue_timeout:
                default: 1s
                description: Expect Continue Timeout. 0 = no timeout.
                format: string
                type: string
                x-go-name: ExpectContinueTimeout
                x-go-tag-default: 1s
                x-go-tag-json: expect_continue_timeout
                x-go-tag-validate: gte=0s
            idle_connection_timeout:
                default: 90s
                description: Idle Connection Timeout. 0 = no timeout.
                format: string
                type: string
                x-go-name: IdleConnTimeout
                x-go-tag-default: 90s
                x-go-tag-json: idle_connection_timeout
                x-go-tag-validate: gte=0s
            key_log_file:
                description: SSL/TLS key log file (useful for debugging)
                type: string
                x-go-name: KeyLogWriter
                x-go-tag-json: key_log_file
            max_conns_per_host:
                default: 0
                description: Max Connections Per Host. 0 = no limit.
                format: int64
                minimum: 0
                type: integer
                x-go-name: MaxConnsPerHost
                x-go-tag-default: "0"
                x-go-tag-json: max_conns_per_host
                x-go-tag-validate: gte=0
            max_idle_connections:
                default: 100
                description: Max Idle Connections. 0 = no limit.
                format: int64
                minimum: 0
                type: integer
                x-go-name: MaxIdleConns
                x-go-tag-default: "100"
                x-go-tag-json: max_idle_connections
                x-go-tag-validate: gte=0
            max_idle_connections_per_host:
                default: 5
                description: Max Idle Connections per host. 0 = no limit.
                format: int64
                minimum: 0
                type: integer
                x-go-name: MaxIdleConnsPerHost
                x-go-tag-default: "5"
                x-go-tag-json: max_idle_connections_per_host
                x-go-tag-validate: gte=0
            max_response_header_bytes:
                default: 0
                description: Max Response Header Bytes. 0 = no limit.
                format: int64
                minimum: 0
                type: integer
                x-go-name: MaxResponseHeaderBytes
                x-go-tag-default: "0"
                x-go-tag-json: max_response_header_bytes
                x-go-tag-validate: gte=0
            network_keep_alive:
                default: 30s
                description: Network level keep-alive duration
                format: string
                type: string
                x-go-name: NetworkKeepAlive
                x-go-tag-default: 30s
                x-go-tag-json: network_keep_alive
                x-go-tag-validate: gte=0s
            network_timeout:
                default: 30s
                description: Timeout for making network connection
                format: string
                type: string
                x-go-name: NetworkTimeout
                x-go-tag-default: 30s
                x-go-tag-json: network_timeout
                x-go-tag-validate: gte=0s
            proxy_connect_header:
                $ref: '#/definitions/Header'
                description: Proxy Connect Header - `map[string][]string`
                x-go-tag-json: proxy_connect_header,omitempty
                x-go-tag-validate: omitempty
            read_buffer_size:
                default: 0
                description: Read Buffer Size. 0 = 4 KB
                format: int64
                minimum: 0
                type: integer
                x-go-name: ReadBufferSize
                x-go-tag-default: "0"
                x-go-tag-json: read_buffer_size
                x-go-tag-validate: gte=0
            response_header_timeout:
                default: 0s
                description: Response Header Timeout. 0 = no timeout.
                format: string
                type: string
                x-go-name: ResponseHeaderTimeout
                x-go-tag-default: 0s
                x-go-tag-json: response_header_timeout
                x-go-tag-validate: gte=0s
            timeout:
                default: 60s
                description: HTTP client timeout - Timeouts include connection time, redirects, reading the response and so on. 0 = no timeout.
                format: string
                type: string
                x-go-name: Timeout
                x-go-tag-default: 60s
                x-go-tag-json: timeout
                x-go-tag-validate: gte=0s
            tls:
                $ref: '#/definitions/ClientTLSConfig'
                description: Client TLS configuration
                x-go-tag-json: tls
            tls_handshake_timeout:
                default: 10s
                description: TLS Handshake Timeout. 0 = no timeout
                format: string
                type: string
                x-go-name: TLSHandshakeTimeout
                x-go-tag-default: 10s
                x-go-tag-json: tls_handshake_timeout
                x-go-tag-validate: gte=0s
            use_proxy:
                default: false
                description: Use Proxy
                type: boolean
                x-go-name: UseProxy
                x-go-tag-default: "false"
                x-go-tag-json: use_proxy
            write_buffer_size:
                default: 0
                description: Write Buffer Size. 0 = 4 KB.
                format: int64
                minimum: 0
                type: integer
                x-go-name: WriteBufferSize
                x-go-tag-default: "0"
                x-go-tag-json: write_buffer_size
                x-go-tag-validate: gte=0
        title: HTTPClientConfig holds configuration for HTTP Client.
        type: object
        x-go-package: github.com/fluxninja/aperture/v2/pkg/net/http
...
    Header:
        additionalProperties:
            items:
                type: string
            type: array
        description: |-
            The keys should be in canonical form, as returned by
            CanonicalHeaderKey.
        title: A Header represents the key-value pairs in an HTTP header.
        type: object
        x-go-package: net/http
...
    PrometheusConfig:
        properties:
            address:
                description: Address of the Prometheus server
                pattern: ((((^$)|(^([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9]):[0-9]+$))|(^https?://[a-zA-Z0-9\-\.]+\.[a-zA-Z]{2,3}(:[a-zA-Z0-9]*)?/?([a-zA-Z0-9\-\._\?\,\'/\\\+&amp;%\$#\=~])*$))|(^([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])$))|(^$)
                type: string
                x-go-name: Address
                x-go-tag-json: address,omitempty
                x-go-tag-validate: omitempty,hostname_port|url|fqdn,omitempty
                x-pattern-rules: empty | hostname_port | url | fqdn | empty
            labels:
                description: A list of labels to be attached to every query
                items:
                    $ref: '#/definitions/PrometheusLabel'
                type: array
                x-go-name: Labels
                x-go-tag-json: labels,omitempty
        title: PrometheusConfig holds configuration for Prometheus Server.
        type: object
        x-go-package: github.com/fluxninja/aperture/v2/pkg/prometheus/config
    PrometheusLabel:
        description: |-
            PrometheusLabel holds Name->Value mapping for the label that will
            be attached to every PromQL query executed by the controller.
        properties:
            name:
                type: string
                x-go-name: Name
                x-go-tag-json: name
            value:
                type: string
                x-go-name: Value
                x-go-tag-json: value
        type: object
        x-go-package: github.com/fluxninja/aperture/v2/pkg/prometheus/config
...
info:
    description: Aperture Agent
    title: Agent
paths:
...
    /prometheus:
        post:
            operationId: Prometheus
            parameters:
                - in: body
                  schema:
                    $ref: '#/definitions/PrometheusConfig'
                - in: body
                  name: http_client
                  schema:
                    $ref: '#/definitions/HTTPClientConfig'
            tags:
                - common-configuration
            x-fn-config-env: true
...
swagger: "2.0"